Added run_unit_tests.sh micro-harness and tdycore_tests.h header.
This shell script queries a unit test exe for its number of tests and
its appopriate number of MPI processes, and then runs those tests one
at a time. If any test fails, it exits and returns the status code with
an error message.
