+++ src/tests/test_tdyinit.c
@@ -1,11 +1,4 @@
-#include <stdarg.h>
-#include <stddef.h>
-#include <setjmp.h>
-#include <string.h>
-#include <stdlib.h>
-#include <stdbool.h>
-#include <cmocka.h>
-
+#include <tdycore_tests.h>
 #include <tdycore.h>
 
 // This unit test suite tests the initialization and finalization of the
@@ -14,9 +7,6 @@
 // Globals for holding command line arguments.
 static int argc_;
 static char **argv_;
-
-// Flag indicating whether tests are run in their own process.
-static bool isolate_tests = false;
 
 // Test whether TDyInit works and initializes MPI properly.
 static void TestTDyInit(void **state)
@@ -28,17 +18,13 @@ static void TestTDyInit(void **state)
   MPI_Initialized(&mpi_initialized);
   assert_true(mpi_initialized);
 
-  if (isolate_tests) {
-    TDyFinalize();
-  }
+  TDyFinalize();
 }
 
 // Test whether the PETSC_COMM_WORLD communicator behaves properly within cmocka.
 static void TestPetscCommWorld(void **state)
 {
-  if (isolate_tests) {
-    TDyInit(argc_, argv_);
-  }
+  TDyInit(argc_, argv_);
 
   int num_procs;
   MPI_Comm_size(PETSC_COMM_WORLD, &num_procs);
@@ -49,17 +35,13 @@ static void TestPetscCommWorld(void **state)
   assert_true(rank >= 0);
   assert_true(rank < num_procs);
 
-  if (isolate_tests) {
-    TDyFinalize();
-  }
+  TDyFinalize();
 }
 
 // Test whether MPI performs as expected within CMocka's environment.
 static void TestMPIAllreduce(void **state)
 {
-  if (isolate_tests) {
-    TDyInit(argc_, argv_);
-  }
+  TDyInit(argc_, argv_);
 
   // Let's see if we can properly sum over all ranks.
   int num_procs;
@@ -75,42 +57,6 @@ static void TestMPIAllreduce(void **state)
   TDyFinalize();
 }
 
-static int _run_selected_tests(const char* command,
-                               int num_tests,
-                               const struct CMUnitTest tests[num_tests]) {
-  // Set our test isolation flag.
-  isolate_tests = true;
-
-  if (strcasecmp(command, "count") == 0) { // asked for # of tests
-    fprintf(stdout, "%d\n", num_tests);
-    return 0;
-  } else { // asked for a specific test index
-    // Try to interpret the argument as an index for the desired test.
-    char *endptr;
-    long index = strtol(command, &endptr, 10);
-    if (*endptr == '\0') { // got a valid index!
-      if ((index < 0) || (index >= num_tests)) {
-        fprintf(stderr, "Invalid test index: %ld (must be in [0, %d])\n",
-            index, num_tests);
-        return 1;
-      } else {
-        const struct CMUnitTest selected_tests[] = { tests[index] };
-        return cmocka_run_group_tests(selected_tests, NULL, NULL);
-      }
-    } else {
-      fprintf(stderr, "Invalid command: %s (must be 'count' or index)\n",
-          command);
-      return 1;
-    }
-  }
-}
-
-#define run_selected_tests(argc, argv, tests) \
-  (argc > 1) ? _run_selected_tests(argv[1], \
-                                   sizeof(tests) / sizeof((tests)[0]), \
-                                   tests) \
-             : cmocka_run_group_tests(tests, NULL, NULL)
-
 int main(int argc, char* argv[])
 {
   // Stash command line arguments.
@@ -122,8 +68,8 @@ int main(int argc, char* argv[])
   {
     cmocka_unit_test(TestTDyInit),
     cmocka_unit_test(TestPetscCommWorld),
-    cmocka_unit_test(TestMPIAllreduce)
+    cmocka_unit_test(TestMPIAllreduce),
   };
 
-  run_selected_tests(argc, argv, tests);
+  run_selected_tests(argc, argv, tests, 2);
 }
